[chore #157956705] Move tests to test_auth and resolve any conflicts
